+++ src/hooks/useContentType.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback, useMemo } from 'react';
+import { useState, useEffect } from 'react';
 import { DEFAULT_CONTENT_TYPE, DEFAULT_CONTENT_TYPE_NAME } from '../constants/ContentType';
 import { Settings } from '../dashboardWebView/models';
 import { ContentType, PanelSettings } from '../models';
@@ -11,7 +11,22 @@ export default function useContentType(
 
   useEffect(() => {
     if (settings) {
-      const contentTypeName = (metadata.type as string) || DEFAULT_CONTENT_TYPE_NAME;
+      let contentTypeName = DEFAULT_CONTENT_TYPE_NAME;
+
+      if (metadata?.type) {
+        contentTypeName = metadata.type;
+      }
+
+      // Get the content type by the folder name
+      const pageFolders = settings.contentFolders;
+      let pageFolderMatches = pageFolders.filter((folder) => metadata.filePath.includes(folder.path));
+
+      // Sort by longest path
+      pageFolderMatches = pageFolderMatches.sort((a, b) => b.path.length - a.path.length);
+      if (pageFolderMatches.length > 0 && pageFolderMatches[0].contentTypes && pageFolderMatches[0].contentTypes.length === 1) {
+        contentTypeName = pageFolderMatches[0].contentTypes[0];
+      }
+
       let ct = settings.contentTypes.find((ct) => ct.name === contentTypeName);
 
       if (!ct) {