+++ CHANGELOG.md
@@ -18,7 +18,9 @@
 - [#633](https://github.com/estruyf/vscode-front-matter/issues/633): Fix custom taxonomy used as single value
 - [#634](https://github.com/estruyf/vscode-front-matter/issues/634): Fix public folder processing in case of Hexo
 - [#635](https://github.com/estruyf/vscode-front-matter/issues/635): Fix UI themes make some links hard to discover or see
-- [#636](https://github.com/estruyf/vscode-front-matter/issues/636): Fix UI themes make some links hard to discover or see
+- [#636](https://github.com/estruyf/vscode-front-matter/issues/636): Fix media browser lightbox background color
+- [#641](https://github.com/estruyf/vscode-front-matter/issues/641): Fix file parsing of the localization file
+
 
 ## [9.0.0] - 2023-08-21 - [Release notes](https://beta.frontmatter.codes/updates/v9.0.0)