+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## [9.1.0] - 2023-08-31
+
+### ✨ New features
+
+- [#630](https://github.com/estruyf/vscode-front-matter/issues/630): Italian localization added (thanks to [Emanuele Bartolesi](https://github.com/kasuken))
+
+### ⚡️ Optimizations
+
+- [#639](https://github.com/estruyf/vscode-front-matter/issues/639): Add check for content type in page folders setting
+- [#642](https://github.com/estruyf/vscode-front-matter/issues/642): Store the search index in a storage file
+
+### 🐞 Fixes
+
+- [#628](https://github.com/estruyf/vscode-front-matter/issues/628): Fix path argument and JSON data on custom scripts
+- [#633](https://github.com/estruyf/vscode-front-matter/issues/633): Fix custom taxonomy used as single value
+- [#634](https://github.com/estruyf/vscode-front-matter/issues/634): Fix public folder processing in case of Hexo
+- [#635](https://github.com/estruyf/vscode-front-matter/issues/635): Fix UI themes make some links hard to discover or see
+- [#636](https://github.com/estruyf/vscode-front-matter/issues/636): Fix media browser lightbox background color
+- [#641](https://github.com/estruyf/vscode-front-matter/issues/641): Fix file parsing of the localization file
+
+
 ## [9.0.0] - 2023-08-21 - [Release notes](https://beta.frontmatter.codes/updates/v9.0.0)
 
 ### 🌎 Multilingual support